blockchain_interface/requirements.txt
substrate-interface>=1.1.2
websocket-client>=1.3.3
pytest>=7.0.0
pytest-mock>=3.10.0
